Change Int to Long.

+++ src/main/kotlin/edu/agh/susgame/dto/socket/ServerSocketMessage.kt
@@ -36,7 +36,7 @@ sealed class ServerSocketMessage {
         val edges: List<EdgeDTO>,
         val players: List<PlayerDTO>,
         val gameStatus: GameStatus,
-        val timeLeft: Int, // time in seconds left
+        val timeLeft: Long, // time in seconds left
     ) : ServerSocketMessage()
 
     /**
